requirements.txt
django==2.0.2
djangorestframework==3.7.7
django-cors-headers==2.1.0
django-pygments==0.1
django-markdownx==2.0.22
markdown2==2.3.5
gunicorn==19.6.0
